requirements.txt
# RAG Second Brain v16 - Requirements
# Core ML
torch>=2.0.0
transformers>=4.36.0
sentence-transformers>=2.2.0

# Vector search
faiss-cpu>=1.7.4

# Knowledge Graph & Ontology
owlready2>=0.45
networkx>=3.0
rdflib>=7.0.0

# NLP & Entity Linking
spacy>=3.7.0

# Data & Computation
scipy>=1.11.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Datasets
datasets>=2.14.0

# Experiment tracking
wandb>=0.15.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Evaluation
rouge-score>=0.1.2

# Development
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
